inform/urls.py
from articles.feeds import ArticleFeed
from django.conf import settings
from django.conf.urls import url
from django.contrib import admin
from django.urls import include, path
from django.views.decorators.csrf import csrf_exempt
from graphene_django.views import GraphQLView
from grapple import urls as grapple_urls
from wagtail_transfer import urls as wagtailtransfer_urls

from wagtail.admin import urls as wagtailadmin_urls
from wagtail.core import urls as wagtail_urls
from wagtail.contrib.sitemaps.views import sitemap
from wagtail.documents import urls as wagtaildocs_urls


def fail(_request):
    raise Exception("Controlled fail")


urlpatterns = [
    path("api/graphql/",csrf_exempt(GraphQLView.as_view()),name="grapple_graphql"),
    path("api/", include(grapple_urls)),
    path("admin/", include(wagtailadmin_urls))
]


if settings.DEBUG:
    import debug_toolbar
    from django.conf.urls.static import static
    from django.contrib.staticfiles.urls import staticfiles_urlpatterns

    # Serve static and media files from development server
    urlpatterns += staticfiles_urlpatterns()
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

urlpatterns = urlpatterns + [
    # For anything not caught by a more specific rule above, hand over to
    # Wagtail's page serving mechanism. This should be the last pattern in
    # the list:
    path("", include(wagtail_urls)),
    # Alternatively, if you want Wagtail pages to be served from a subpath
    # of your site, rather than the site root:
    #    path("pages/", include(wagtail_urls)),
]

admin.site.site_header = f"Inform CMS v{settings.VERSION}"
