requirements.txt
aniso8601==7.0.0 ; python_version >= "3.9" and python_version < "4.0"
anyascii==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.5.0 ; python_version >= "3.9" and python_version < "4.0"
arrow==1.2.2 ; python_version >= "3.9" and python_version < "4.0"
asgiref==3.5.0 ; python_version >= "3.9" and python_version < "4.0"
backoff==1.11.1 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.9.3 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.22.0 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.25.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2021.10.8 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==2.0.12 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.4 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
django-debug-toolbar==3.2.4 ; python_version >= "3.9" and python_version < "4.0"
django-filter==21.1 ; python_version >= "3.9" and python_version < "4.0"
django-graphiql-debug-toolbar @ git+https://github.com/Chihiro-Software/django-grapple-graphiql-debug-toolbar@main ; python_version >= "3.9" and python_version < "4.0"
django-health-check==3.16.5 ; python_version >= "3.9" and python_version < "4.0"
django-iam-dbauth==0.1.4 ; python_version >= "3.9" and python_version < "4.0"
django-modelcluster==5.3 ; python_version >= "3.9" and python_version < "4.0"
django-storages==1.12.3 ; python_version >= "3.9" and python_version < "4.0"
django-taggit==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
django-treebeard==4.5.1 ; python_version >= "3.9" and python_version < "4.0"
django-wagtail-feeds==0.1.0 ; python_version >= "3.9" and python_version < "4.0"
django==3.2.13 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.13.1 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
draftjs-exporter==2.1.7 ; python_version >= "3.9" and python_version < "4.0"
elasticsearch-dsl==7.4.0 ; python_version >= "3.9" and python_version < "4.0"
elasticsearch==7.13.4 ; python_version >= "3.9" and python_version < "4"
et-xmlfile==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
graphene-django==2.13.0 ; python_version >= "3.9" and python_version < "4.0"
graphene==2.1.9 ; python_version >= "3.9" and python_version < "4.0"
graphql-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
graphql-relay==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
h11==0.12.0 ; python_version >= "3.9" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.14.7 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
humanize==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.3 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
l18n==2021.3 ; python_version >= "3.9" and python_version < "4.0"
lxml==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
openpyxl==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
phpserialize==1.3 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.1.0 ; python_version >= "3.9" and python_version < "4.0"
promise==2.3 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.9.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2022.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
requests-aws4auth==1.1.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.27.1 ; python_version >= "3.9" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
rx==1.6.1 ; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.5.2 ; python_version >= "3.9" and python_version < "4.0"
sentry-sdk==1.5.10 ; python_version >= "3.9" and python_version < "4.0"
shortcodes==5.4.0 ; python_version >= "3.9" and python_version < "4.0"
singledispatch==3.7.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.3.2.post1 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.2 ; python_version >= "3.9" and python_version < "4.0"
tablib[xls,xlsx]==3.2.1 ; python_version >= "3.9" and python_version < "4.0"
telepath==0.2 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.64.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.2.0 ; python_version >= "3.9" and python_version < "4.0"
unidecode==1.3.4 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.9 ; python_version >= "3.9" and python_version < "4"
wagtail-color-panel==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
wagtail-grapple @ git+https://github.com/przemub/wagtail-grapple-flyte.git@feature/plural_query_field_limit ; python_version >= "3.9" and python_version < "4.0"
wagtail-headless-preview==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
wagtail-icon-picker @ git+https://github.com/przemub/wagtail-icon-picker.git@tmf-release ; python_version >= "3.9" and python_version < "4.0"
wagtail-orderable==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
wagtail-transfer @ git+https://gitlab.com/ggmr/wagtail-transfer.git@custom-release-1 ; python_version >= "3.9" and python_version < "4.0"
wagtail==2.16.2 ; python_version >= "3.9" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
whitenoise==6.0.0 ; python_version >= "3.9" and python_version < "4.0"
willow==1.4.1 ; python_version >= "3.9" and python_version < "4.0"
xlrd==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
xlsxwriter==3.0.3 ; python_version >= "3.9" and python_version < "4.0"
xlwt==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
